rag-chatbot/requirements.txt
flask==2.0.1
werkzeug==2.0.3
pymongo==4.3.3
langchain==0.0.267
sentence-transformers==2.2.2
langchain-community==0.0.10
python-dotenv==0.19.0
PyPDF2==3.0.1
pytest==7.3.1